fix/SV-107: 대소문자 naming 수정
