+++ src/Router.tsx
@@ -3,7 +3,7 @@ import Main from '@pages/main';
 import { Suspense } from 'react';
 import { Route, Routes } from 'react-router-dom';
 
-import Layout from './common/components/Layout';
+import Layout from './common/components/Layoutaaa';
 
 // react-router-dom v7에 관한 Future Flag 경고창 무시
 const originalWarn = console.warn;
